+++ cmd/entrypoint/watcher.go
@@ -3,314 +3,287 @@ package entrypoint
 import (
 	"context"
 	"encoding/json"
-	"fmt"
 	"io/ioutil"
-	"log"
-	"os"
-	"strings"
 	"sync/atomic"
+	"time"
 
 	"github.com/datawire/ambassador/pkg/acp"
 	"github.com/datawire/ambassador/pkg/debug"
 	"github.com/datawire/ambassador/pkg/kates"
-	snapshotTypes "github.com/datawire/ambassador/pkg/snapshot/v1"
+	"github.com/datawire/ambassador/pkg/snapshot/v1"
 	"github.com/datawire/ambassador/pkg/watt"
 	"github.com/datawire/dlib/dlog"
 )
 
-// thingToWatch is... uh... a thing we're gonna watch. Specifically, it's a
-// K8s type name and an optional field selector.
-type thingToWatch struct {
-	typename      string
-	fieldselector string
-}
-
-// watcher is the thing that watches all the K8s stuff we're interested in.
 func watcher(ctx context.Context, ambwatch *acp.AmbassadorWatcher, encoded *atomic.Value) {
-	crdYAML, err := ioutil.ReadFile(findCRDFilename())
-	if err != nil {
-		panic(err)
-	}
-
-	crdObjs, err := kates.ParseManifests(string(crdYAML))
-	if err != nil {
-		panic(err)
-	}
-	validator, err := kates.NewValidator(nil, crdObjs)
-	if err != nil {
-		panic(err)
-	}
-
 	client, err := kates.NewClient(kates.ClientConfig{})
 	if err != nil {
 		panic(err)
-	}
-
-	ns := kates.NamespaceAll
-	if IsAmbassadorSingleNamespace() {
-		ns = GetAmbassadorNamespace()
-	}
-
-	fs := GetAmbassadorFieldSelector()
-	ls := GetAmbassadorLabelSelector()
-
-	endpointFs := "metadata.namespace!=kube-system"
-	if fs != "" {
-		endpointFs += fmt.Sprintf(",%s", fs)
 	}
 
 	serverTypeList, err := client.ServerPreferredResources()
 	if err != nil {
 		// It's possible that an error prevented listing some apigroups, but not all; so
 		// process the output even if there is an error.
-		log.Printf("Warning, unable to list api-resources: %v", err)
-	}
-	serverTypes := make(map[string]kates.APIResource, len(serverTypeList))
-	for _, typeinfo := range serverTypeList {
-		serverTypes[typeinfo.Name+"."+typeinfo.Group] = typeinfo
+		dlog.Infof(ctx, "Warning, unable to list api-resources: %v", err)
 	}
 
-	// We set interestingTypes to the list of types that we'd like to watch (if that type exits
-	// in this cluster).
-	//
-	// - The key in the map is the how we'll label them in the snapshot we pass to the rest of
-	//   Ambassador.
-	// - The typename in the map values should be the qualified "${name}.${group}", where
-	//   "${name} is lowercase+plural.
-	// - If the map value doesn't set a field selector, then `fs` (above) will be used.
-	//
-	// Most of the interestingTypes are static, but it's completely OK to add types based
-	// on runtime considerations, as we do for IngressClass and the KNative stuff.
-	interestingTypes := map[string]thingToWatch{
-		"Services":   {typename: "services."},
-		"K8sSecrets": {typename: "secrets."}, // Note: "K8sSecrets" is not the "obvious" keyname
-		"Endpoints":  {typename: "endpoints.", fieldselector: endpointFs},
+	interestingTypes := GetInterestingTypes(ctx, serverTypeList)
+	queries := GetQueries(ctx, interestingTypes)
 
-		//"Ingresses": {typename: "ingresses.networking.k8s.io"}, // new in Kubernetes 1.14, deprecating ingresses.extensions
-		"Ingresses": {typename: "ingresses.extensions"}, // new in Kubernetes 1.2
+	// **** SETUP DONE for the Kubernetes Watcher
 
-		"AuthServices":                {typename: "authservices.getambassador.io"},
-		"ConsulResolvers":             {typename: "consulresolvers.getambassador.io"},
-		"DevPortals":                  {typename: "devportals.getambassador.io"},
-		"Hosts":                       {typename: "hosts.getambassador.io"},
-		"KubernetesEndpointResolvers": {typename: "kubernetesendpointresolvers.getambassador.io"},
-		"KubernetesServiceResolvers":  {typename: "kubernetesserviceresolvers.getambassador.io"},
-		"LogServices":                 {typename: "logservices.getambassador.io"},
-		"Mappings":                    {typename: "mappings.getambassador.io"},
-		"Modules":                     {typename: "modules.getambassador.io"},
-		"RateLimitServices":           {typename: "ratelimitservices.getambassador.io"},
-		"TCPMappings":                 {typename: "tcpmappings.getambassador.io"},
-		"TLSContexts":                 {typename: "tlscontexts.getambassador.io"},
-		"TracingServices":             {typename: "tracingservices.getambassador.io"},
-	}
-
-	if !IsAmbassadorSingleNamespace() {
-		interestingTypes["IngressClasses"] = thingToWatch{typename: "ingressclasses.networking.k8s.io"} // new in Kubernetes 1.18
-	}
-
-	if IsKnativeEnabled() {
-		// Note: These keynames have a "KNative" prefix, to avoid clashing with the
-		// standard "networking.k8s.io" and "extensions" types.
-		interestingTypes["KNativeClusterIngresses"] = thingToWatch{typename: "clusteringresses.networking.internal.knative.dev"}
-		interestingTypes["KNativeIngresses"] = thingToWatch{typename: "ingresses.networking.internal.knative.dev"}
-	}
-
-	var queries []kates.Query
-	for snapshotname, queryinfo := range interestingTypes {
-		if _, haveType := serverTypes[queryinfo.typename]; !haveType {
-			log.Printf("Warning, unable to watch %s, unknown kind.", queryinfo.typename)
-			continue
-		}
-
-		query := kates.Query{
-			Namespace:     ns,
-			Name:          snapshotname,
-			Kind:          queryinfo.typename,
-			FieldSelector: queryinfo.fieldselector,
-			LabelSelector: ls,
-		}
-		if query.FieldSelector == "" {
-			query.FieldSelector = fs
-		}
-
-		queries = append(queries, query)
-	}
-
-	snapshot := NewKubernetesSnapshot()
-	acc := client.Watch(ctx, queries...)
-
-	consulSnapshot := &watt.ConsulSnapshot{}
-	consul := newConsul(ctx, &consulWatcher{})
-
-	// Time to fire up the stuff we need to watch the filesystem for Istio
-	// certs -- specifically, we need an FSWatcher to watch the filesystem,
-	// an IstioCert to manage the cert, and an update channel to hear about
-	// new Istio stuff.
-	//
-	// The actual functionality here is currently keyed off the environment
-	// variable AMBASSADOR_ISTIO_SECRET_DIR, but we set the update channel
-	// either way to keep the select logic below simpler. If the environment
-	// variable is unset, we never instantiate the FSWatcher or IstioCert,
-	// so there will never be any updates on the update channel.
-	istioCertUpdateChannel := make(chan IstioCertUpdate)
-
-	// OK. Are we supposed to watch anything?
-	secretDir := os.Getenv("AMBASSADOR_ISTIO_SECRET_DIR")
-
-	if secretDir != "" {
-		// Yup, get to it. First, fire up the IstioCert, and tell it to
-		// post to our update channel from above.
-		icert := NewIstioCert(secretDir, "istio-certs", GetAmbassadorNamespace(), istioCertUpdateChannel)
-
-		// Next up, fire up the FSWatcher...
-		fsw, err := NewFSWatcher(ctx)
-
-		if err != nil {
-			// Really, this should never, ever happen.
-			panic(err)
-		}
-
-		// ...then tell the FSWatcher to watch the Istio cert directory,
-		// and give it a handler function that'll update the IstioCert
-		// in turn.
-		//
-		// XXX This handler function is really just an impedance matcher.
-		// Maybe IstioCert should just have a "HandleFSWEvent"...
-		fsw.WatchDir(ctx, secretDir,
-			func(ctx context.Context, event FSWEvent) {
-				// Is this a deletion?
-				deleted := (event.Op == FSWDelete)
-
-				// OK. Feed this event into the IstioCert.
-				icert.HandleEvent(ctx, event.Path, deleted)
-			},
-		)
-	}
-
-	var unsentDeltas []*kates.Delta
-
-	invalid := map[string]*kates.Unstructured{}
-	isValid := func(un *kates.Unstructured) bool {
-		key := string(un.GetUID())
-		err := validator.Validate(ctx, un)
-		if err != nil {
-			copy := un.DeepCopy()
-			copy.Object["errors"] = err.Error()
-			invalid[key] = copy
-			return false
-		} else {
-			delete(invalid, key)
-			return true
+	notify := func(ctx context.Context, disposition SnapshotDisposition, snap *snapshot.Snapshot) {
+		if disposition == SnapshotReady {
+			notifyReconfigWebhooks(ctx, ambwatch)
 		}
 	}
 
+	k8sSrc := newK8sSource(client)
+	consulSrc := &consulWatcher{}
+	istioCertSrc := newIstioCertSource()
+
+	watcherLoop(ctx, encoded, k8sSrc, queries, consulSrc, istioCertSrc, notify)
+}
+
+type SnapshotProcessor func(context.Context, SnapshotDisposition, *snapshot.Snapshot)
+type SnapshotDisposition int
+
+const (
+	// Indicates the watcher is still in the booting process and the snapshot has dangling pointers.
+	SnapshotIncomplete SnapshotDisposition = iota
+	// Indicates that the watcher is deferring processing of the snapshot because it is considered
+	// to be a product of churn.
+	SnapshotDefer
+	// Indicates that the watcher is dropping the snapshot because it has determined that it is
+	// logically a noop.
+	SnapshotDrop
+	// Indicates that the snapshot is ready to be processed.
+	SnapshotReady
+)
+
+// watcher is _the_ thing that watches all the different kinds of Ambassador configuration
+// events that we care about. This right here is pretty much the root of everything flowing
+// into Ambassador from the outside world, so:
+//
+// ******** READ THE FOLLOWING COMMENT CAREFULLY! ********
+//
+// Since this is where _all_ the different kinds of these events (K8s, Consul, filesystem,
+// whatever) are brought together and examined, and where we pass judgement on whether or
+// not a given event is worth reconfiguring Ambassador or not, the interactions between
+// this function and other pieces of the system can be quite a bit more complex than you
+// might expect. There are two really huge things you should be bearing in mind if you
+// need to work on this:
+//
+// 1. The set of things we're watching is not static, but it must converge.
+//
+//    An example: you can set up a Kubernetes watch that finds a KubernetesConsulResolver
+//    resource, which will then prompt a new Consul watch to happen. At present, nothing
+//    that that Consul watch could find is capable of prompting a new Kubernetes watch to
+//    be created. This is important: it would be fairly easy to change things such that
+//    there is a feedback loop where the set of things we watch does not converge on a
+//    stable set. If such a loop exists, fixing it will probably require grokking this
+//    watcher function, kates.Accumulator, and maybe the reconcilers in consul.go and
+//    endpoints.go as well.
+//
+// 2. No one source of input events can be allowed to alter the event stream for another
+//    source.
+//
+//    An example: at one point, a bug in the watcher function resulted in the Kubernetes
+//    watcher being able to decide to short-circuit a watcher iteration -- which had the
+//    effect of allowing the K8s watcher to cause _Consul_ events to be ignored. That's
+//    not OK. To guard against this:
+//
+//    A. Refrain from adding state to the watcher loop.
+//    B. Try very very hard to keep logic that applies to a single source within that
+//       source's specific case in the watcher's select statement.
+//    C. Don't add any more select statements, so that B. above is unambiguous.
+//
+// 3. If you add a new channel to watch, you MUST make sure it has a way to let the loop
+//    know whether it saw real changes, so that the short-circuit logic works correctly.
+//    That said, recognize that the way it works now, with the state for the individual
+//    watchers in the watcher() function itself is a crock, and the path forward is to
+//    refactor them into classes that can separate things more cleanly.
+//
+// 4. If you don't fully understand everything above, _do not touch this function without
+//    guidance_.
+func watcherLoop(ctx context.Context, encoded *atomic.Value, k8sSrc K8sSource, queries []kates.Query,
+	consulWatcher Watcher, istioCertSrc IstioCertSource, notify SnapshotProcessor) {
+	// These timers keep track of various parts of the processing of the watcher loop. They don't
+	// directly impact the logic at all.
 	dbg := debug.FromContext(ctx)
 
 	katesUpdateTimer := dbg.Timer("katesUpdate")
 	consulUpdateTimer := dbg.Timer("consulUpdate")
+	istioCertUpdateTimer := dbg.Timer("istioCertUpdate")
 	notifyWebhooksTimer := dbg.Timer("notifyWebhooks")
 	parseAnnotationsTimer := dbg.Timer("parseAnnotations")
 	reconcileSecretsTimer := dbg.Timer("reconcileSecrets")
 	reconcileConsulTimer := dbg.Timer("reconcileConsul")
+	reconcileEndpointsTimer := dbg.Timer("reconcileEndpoints")
 
+	// Synthesize the low(ish)-level Kubernetes watcher, then use it to synthesize
+	// the Kubernetes watch manager.
+	k8sWatcher := k8sSrc.Watch(ctx, queries...)
+	k8s := newK8sWatchManager(ctx, k8sWatcher)
+	validator := newResourceValidator()
+
+	// Likewise for the Istio cert watcher and manager.
+	istioCertWatcher := istioCertSrc.Watch(ctx)
+	istio := newIstioCertWatchManager(ctx, istioCertWatcher)
+
+	consul := newConsul(ctx, consulWatcher) // Consul Watcher: state manager
+
+	// **** STATE for the Consul Watcher.
+	//
+	// To track Consul things, we again need a (different kind of) snapshot
+	// and a "consul" object. The snapshot, again, is our view of the stuff
+	// in the Consul world that applies to us; the consul object doesn't so
+	// much have to manage consistency as it has to manage what we tell Consul
+	// we're interested in.
+	consulSnapshot := &watt.ConsulSnapshot{} // Consul Watcher: core state
+	// XXX Temporary hack: give the Consul watcher a trivial way to note that
+	// it saw changes.  This is important because if any of the watchers see
+	// changes, we can't short-circuit the reconfiguration.
+	consulChangesPresent := false // Consul Watcher: core state
+
+	// **** STATE (again) for the Kubernetes Watcher
+	//
+	// We use kates.Delta objects to indicate to the rest of Ambassador
+	// what has actually changed between one snapshot and the next.
+	// unsentDeltas buffers deltas across iterations if a non-bootstrapped
+	// watcher short-circuits, while k8s.deltas is just the current deltas
+	// for the Kubernetes watcher during a given iteration.
+	var unsentDeltas []*kates.Delta // K8s Watcher: core state
+
+	// Is this the very first reconfigure we've done?
 	firstReconfig := true
 
 	for {
+		dlog.Debugf(ctx, "WATCHER: --------")
+
+		// XXX Hack: the istioCertWatchManager needs to reset at the start of the
+		// loop, for now. A better way, I think, will be to instead track deltas in
+		// ReconcileSecrets -- that way we can ditch this crap and Istio-cert changes
+		// that somehow don't generate an actual change will still not trigger a
+		// reconfigure.
+		istio.StartLoop(ctx)
+
 		select {
-		case <-acc.Changed():
+		case <-k8s.Changed():
+			// Kubernetes has some changes, so we need to handle them.
 			stop := katesUpdateTimer.Start()
-			var deltas []*kates.Delta
+
 			// We could probably get a win in some scenarios by using this filtered update thing to
 			// pre-exclude based on ambassador-id.
-			if !acc.FilteredUpdate(snapshot, &deltas, isValid) {
-				stop()
+			newChanges := k8s.Update(ctx, func(un *kates.Unstructured) bool {
+				return validator.isValid(ctx, un)
+			})
+			stop()
+
+			if !newChanges {
 				continue
 			}
-			unsentDeltas = append(unsentDeltas, deltas...)
-			stop()
+
 		case <-consul.changed():
+			dlog.Debugf(ctx, "WATCHER: Consul fired")
+
+			// Consul has some changes. The Consul watcher doesn't currently track
+			// deltas the same way that the K8s watcher does, but OTOH anything we
+			// watch in Consul is something that we know we have a reason to care
+			// about. So we can go ahead and declare that we've seen real Consul
+			// changes here.
+			consulChangesPresent = true
+
 			consulUpdateTimer.Time(func() {
 				consul.update(consulSnapshot)
 			})
-		case icertUpdate := <-istioCertUpdateChannel:
-			// Make a SecretRef for this new secret...
-			ref := snapshotTypes.SecretRef{Name: icertUpdate.Name, Namespace: icertUpdate.Namespace}
 
-			// ...and delete or save, as appropriate.
-			if icertUpdate.Op == "delete" {
-				dlog.Infof(ctx, "IstioCert: certificate %s.%s deleted", icertUpdate.Name, icertUpdate.Namespace)
-				delete(snapshot.FSSecrets, ref)
-			} else {
-				dlog.Infof(ctx, "IstioCert: certificate %s.%s updated", icertUpdate.Name, icertUpdate.Namespace)
-				snapshot.FSSecrets[ref] = icertUpdate.Secret
-			}
-			// Once done here, snapshot.ReconcileSecrets will handle the rest.
+		case icertUpdate := <-istio.Changed():
+			// The Istio cert has some changes, so we need to handle them.
+			istioCertUpdateTimer.Time(func() {
+				istio.Update(ctx, icertUpdate, k8s)
+			})
+
+		// BEFORE ADDING A NEW CHANNEL, READ THE COMMENT AT THE TOP OF THIS
+		// FUNCTION so you don't break the short-circuiting logic.
+
 		case <-ctx.Done():
 			return
 		}
 
 		parseAnnotationsTimer.Time(func() {
-			parseAnnotations(snapshot)
+			parseAnnotations(k8s.snapshot)
 		})
 
 		reconcileSecretsTimer.Time(func() {
-			ReconcileSecrets(snapshot)
+			ReconcileSecrets(k8s.snapshot)
 		})
 		reconcileConsulTimer.Time(func() {
-			ReconcileConsul(ctx, consul, snapshot)
+			ReconcileConsul(ctx, consul, k8s.snapshot)
 		})
 
-		if !consul.isBootstrapped() {
+		reconcileEndpointsTimer.Time(func() {
+			k8s.deltas = ReconcileEndpoints(ctx, k8s.snapshot, k8s.deltas)
+			dlog.Debugf(ctx, "WATCHER: filtered deltas (%d): %s", len(k8s.deltas), deltaSummary(k8s.deltas))
+		})
+
+		// Ugh. We need to remember here whether there were any Kubernetes changes,
+		// because k8s.GetDeltas() will empty the delta set that k8s.UpdatesPresent()
+		// will look at.
+		k8sHasChanges := k8s.UpdatesPresent()
+		unsentDeltas = append(unsentDeltas, k8s.GetDeltas()...)
+
+		sn := &snapshot.Snapshot{
+			Kubernetes: k8s.snapshot,
+			Consul:     consulSnapshot,
+			Invalid:    validator.getInvalid(),
+			Deltas:     unsentDeltas,
+		}
+
+		// Do we have any real changes from any watcher?
+		if !k8sHasChanges && !consulChangesPresent && !istio.UpdatesPresent() {
+			// Nope, no changes at all -- we can short-circuit.
+			dlog.Debugf(ctx, "WATCHER: all deltas filtered out")
+			notify(ctx, SnapshotDrop, sn)
 			continue
 		}
 
-		var invalidSlice []*kates.Unstructured
-		for _, inv := range invalid {
-			invalidSlice = append(invalidSlice, inv)
+		if !consul.isBootstrapped() {
+			notify(ctx, SnapshotIncomplete, sn)
+			continue
 		}
 
-		sn := &snapshotTypes.Snapshot{
-			Kubernetes: snapshot,
-			Consul:     consulSnapshot,
-			Invalid:    invalidSlice,
-			Deltas:     unsentDeltas,
-		}
 		unsentDeltas = nil
 
-		bytes, err := json.MarshalIndent(sn, "", "  ")
+		snapshotJSON, err := json.MarshalIndent(sn, "", "  ")
 		if err != nil {
 			panic(err)
 		}
-		encoded.Store(bytes)
+
+		if envbool("AMBASSADOR_WATCHER_SNAPLOG") {
+			snpath := time.Now().Format("/tmp/20060102T030405-snap.json")
+
+			err = ioutil.WriteFile(snpath, snapshotJSON, 0777)
+
+			if err != nil {
+				dlog.Errorf(ctx, "WATCHER: could not save snapshot to %s: %s", snpath, err)
+			} else {
+				dlog.Debugf(ctx, "WATCHER: saved snapshot as %s", snpath)
+			}
+		}
+
+		// ...then stash this snapshot and fire off webhooks.
+		encoded.Store(snapshotJSON)
 		if firstReconfig {
-			log.Println("Bootstrapped! Computing initial configuration...")
+			dlog.Debugf(ctx, "WATCHER: Bootstrapped! Computing initial configuration...")
 			firstReconfig = false
 		}
+
+		// Finally, use the reconfigure webhooks to let the rest of Ambassador
+		// know about the new configuration.
 		notifyWebhooksTimer.Time(func() {
-			notifyReconfigWebhooks(ctx, ambwatch)
+			notify(ctx, SnapshotReady, sn)
 		})
-
-		// we really only need to be incremental for a subset of things:
-		//  - Mappings & Endpoints are the biggies
-		//  - TLSContext are probably next
-
-		// for Endpoints, we can probably figure out a way to wire things up where we bypass  python entirely:
-		//   - maybe have the python put in a placeholder that the go code fills in
-		//   - maybe use EDS to pump the endpoint data directly to the cluster
 	}
-}
-
-func findCRDFilename() string {
-	searchPath := []string{
-		"/opt/ambassador/etc/crds.yaml",
-		"docs/yaml/ambassador/ambassador-crds.yaml",
-	}
-
-	for _, candidate := range searchPath {
-		if fileExists(candidate) {
-			return candidate
-		}
-	}
-
-	panic(fmt.Sprintf("couldn't find CRDs at any of the following locations: %s", strings.Join(searchPath, ", ")))
 }