+++ python/tests/test_watt_scaling.py
@@ -65,12 +65,7 @@ spec:
         namespace = 'watt-rapid'
 
         # Install Ambassador
-        install_ambassador(namespace=namespace, envs=[
-            {
-                'name': 'AMBASSADOR_SINGLE_NAMESPACE',
-                'value': 'true'
-            }
-        ])
+        install_ambassador(namespace=namespace)
 
         # Install QOTM
         apply_kube_artifacts(namespace=namespace, artifacts=qotm_manifests)